Cast int

+++ robot/src/master.cpp
@@ -126,7 +126,7 @@ void robot::Master::load_webots_to_config()
     for (auto &[id, waypoint] : ast.nodes) {
         waypoint_list.append(Json::objectValue);
         auto &last = waypoint_list[waypoint_list.size() - 1];
-        last["id"] = id;
+        last["id"] = static_cast<int>(id);
         last["x"] = waypoint.translation.x;
         last["y"] = waypoint.translation.z;
         last["type"] = to_string(waypoint.waypointType);
